feat: deprecate "reactElement" support
The second argument for trigger and forward has always been expected
to point to the react element causing the event. It was then provided
as e.target to the event handler. This behaviour encourages exposing
the inner interface of a component which isn't desirable as well as
makes it awkward to use Reon with functional components.